Refactor Clifford symplectic rep lookup and serialization
Addresses a number of Riley's PR suggestions:
- Factors duplicated ExplicitOpModel / ImplicitOpModel symplectic-rep logic
into OpModel.compute_clifford_symplectic_reps().
- Inlines _evaluate_clifford_srep
- Removes the outdated skipped test_construct_with_nonstd_gate_unitary_factory.
- Flattenes state['symplectic_reps'] serialization
- Adds a concrete label-argument remapping example in the clone helper docstring.
- Removes an outdated  @cached_decorator TODO comment.

This commit also fixes a test-discovered bug: map_gate_label_state_space now
must handle LabelTup objects without .time, and was assuming .time always
exists for a label.  Fixed now.
